+++ spring-cloud-starter-tencent-polaris-config/src/main/java/com/tencent/cloud/polaris/config/adapter/PolarisPropertySourceAutoRefresher.java
@@ -18,23 +18,35 @@
 
 package com.tencent.cloud.polaris.config.adapter;
 
+import java.util.Collection;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicBoolean;
 
+import com.tencent.cloud.common.util.JacksonUtils;
 import com.tencent.cloud.polaris.config.config.PolarisConfigProperties;
-import com.tencent.polaris.configuration.api.core.ConfigKVFileChangeEvent;
+import com.tencent.cloud.polaris.config.enums.RefreshType;
+import com.tencent.cloud.polaris.config.spring.property.PlaceholderHelper;
+import com.tencent.cloud.polaris.config.spring.property.SpringValue;
+import com.tencent.cloud.polaris.config.spring.property.SpringValueRegistry;
 import com.tencent.polaris.configuration.api.core.ConfigKVFileChangeListener;
 import com.tencent.polaris.configuration.api.core.ConfigPropertyChangeInfo;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import org.springframework.beans.BeansException;
+import org.springframework.beans.TypeConverter;
+import org.springframework.beans.factory.BeanFactory;
+import org.springframework.beans.factory.BeanFactoryAware;
+import org.springframework.beans.factory.config.ConfigurableBeanFactory;
 import org.springframework.boot.context.event.ApplicationReadyEvent;
+import org.springframework.cloud.context.environment.EnvironmentChangeEvent;
 import org.springframework.cloud.context.refresh.ContextRefresher;
 import org.springframework.context.ApplicationContext;
 import org.springframework.context.ApplicationContextAware;
 import org.springframework.context.ApplicationListener;
+import org.springframework.context.ConfigurableApplicationContext;
+import org.springframework.lang.NonNull;
 import org.springframework.util.CollectionUtils;
 
 /**
@@ -44,7 +56,7 @@ import org.springframework.util.CollectionUtils;
  * @author lepdou 2022-03-28
  */
 public class PolarisPropertySourceAutoRefresher
-		implements ApplicationListener<ApplicationReadyEvent>, ApplicationContextAware {
+		implements ApplicationListener<ApplicationReadyEvent>, ApplicationContextAware, BeanFactoryAware {
 
 	private static final Logger LOGGER = LoggerFactory.getLogger(PolarisPropertySourceAutoRefresher.class);
 
@@ -53,22 +65,35 @@ public class PolarisPropertySourceAutoRefresher
 	private final PolarisPropertySourceManager polarisPropertySourceManager;
 	private final ContextRefresher contextRefresher;
 	private final AtomicBoolean registered = new AtomicBoolean(false);
-	private ApplicationContext applicationContext;
+	private final SpringValueRegistry springValueRegistry;
+	private final PlaceholderHelper placeholderHelper;
+	private ConfigurableApplicationContext context;
+	private TypeConverter typeConverter;
+	private ConfigurableBeanFactory beanFactory;
 
-	public PolarisPropertySourceAutoRefresher(PolarisConfigProperties polarisConfigProperties,
-			PolarisPropertySourceManager polarisPropertySourceManager, ContextRefresher contextRefresher) {
+
+	public PolarisPropertySourceAutoRefresher(
+			PolarisConfigProperties polarisConfigProperties,
+			PolarisPropertySourceManager polarisPropertySourceManager,
+			SpringValueRegistry springValueRegistry,
+			PlaceholderHelper placeholderHelper,
+			ContextRefresher contextRefresher) {
 		this.polarisConfigProperties = polarisConfigProperties;
 		this.polarisPropertySourceManager = polarisPropertySourceManager;
+		this.springValueRegistry = springValueRegistry;
+		this.placeholderHelper = placeholderHelper;
 		this.contextRefresher = contextRefresher;
 	}
 
 	@Override
 	public void setApplicationContext(ApplicationContext applicationContext) throws BeansException {
-		this.applicationContext = applicationContext;
+		this.context = (ConfigurableApplicationContext) applicationContext;
+		this.beanFactory = ((ConfigurableApplicationContext) applicationContext).getBeanFactory();
+		this.typeConverter = this.beanFactory.getTypeConverter();
 	}
 
 	@Override
-	public void onApplicationEvent(ApplicationReadyEvent event) {
+	public void onApplicationEvent(@NonNull ApplicationReadyEvent event) {
 		registerPolarisConfigPublishEvent();
 	}
 
@@ -88,38 +113,105 @@ public class PolarisPropertySourceAutoRefresher
 
 		// register polaris config publish event
 		for (PolarisPropertySource polarisPropertySource : polarisPropertySources) {
-			polarisPropertySource.getConfigKVFile().addChangeListener(new ConfigKVFileChangeListener() {
-				@Override
-				public void onChange(ConfigKVFileChangeEvent configKVFileChangeEvent) {
-					LOGGER.info(
-							"[SCT Config]  received polaris config change event and will refresh spring context."
-									+ " namespace = {}, group = {}, fileName = {}",
-							polarisPropertySource.getNamespace(), polarisPropertySource.getGroup(),
-							polarisPropertySource.getFileName());
+			polarisPropertySource.getConfigKVFile()
+					.addChangeListener((ConfigKVFileChangeListener) configKVFileChangeEvent -> {
 
-					Map<String, Object> source = polarisPropertySource.getSource();
+						LOGGER.info(
+								"[SCT Config] received polaris config change event and will refresh spring context."
+										+ " namespace = {}, group = {}, fileName = {}",
+								polarisPropertySource.getNamespace(),
+								polarisPropertySource.getGroup(),
+								polarisPropertySource.getFileName());
 
-					for (String changedKey : configKVFileChangeEvent.changedKeys()) {
-						ConfigPropertyChangeInfo configPropertyChangeInfo = configKVFileChangeEvent
-								.getChangeInfo(changedKey);
+						Map<String, Object> source = polarisPropertySource.getSource();
 
-						LOGGER.info("[SCT Config] changed property = {}", configPropertyChangeInfo);
+						for (String changedKey : configKVFileChangeEvent.changedKeys()) {
+							ConfigPropertyChangeInfo configPropertyChangeInfo = configKVFileChangeEvent
+									.getChangeInfo(changedKey);
 
-						switch (configPropertyChangeInfo.getChangeType()) {
-						case MODIFIED:
-						case ADDED:
-							source.put(changedKey, configPropertyChangeInfo.getNewValue());
-							break;
-						case DELETED:
-							source.remove(changedKey);
-							break;
+							LOGGER.info("[SCT Config] changed property = {}", configPropertyChangeInfo);
+
+							switch (configPropertyChangeInfo.getChangeType()) {
+							case MODIFIED:
+							case ADDED:
+								source.put(changedKey, configPropertyChangeInfo.getNewValue());
+								break;
+							case DELETED:
+								source.remove(changedKey);
+								break;
+							}
+
+							if (polarisConfigProperties.getRefreshType() == RefreshType.REFLECT) {
+								Collection<SpringValue> targetValues = springValueRegistry.get(beanFactory, changedKey);
+								if (targetValues == null || targetValues.isEmpty()) {
+									continue;
+								}
+								// update the attribute with @Value annotation
+								for (SpringValue val : targetValues) {
+									updateSpringValue(val);
+								}
+							}
 						}
-					}
 
-					// rebuild beans with @RefreshScope annotation
-					contextRefresher.refresh();
-				}
-			});
+						if (polarisConfigProperties.getRefreshType() == RefreshType.REFLECT) {
+							// update @ConfigurationProperties beans
+							context.publishEvent(new EnvironmentChangeEvent(context, configKVFileChangeEvent.changedKeys()));
+						}
+						else {
+							contextRefresher.refresh();
+						}
+					});
 		}
+	}
+
+	private void updateSpringValue(SpringValue springValue) {
+		try {
+			Object value = resolvePropertyValue(springValue);
+			springValue.update(value);
+
+			LOGGER.info("Auto update polaris changed value successfully, new value: {}, {}", value,
+					springValue);
+		}
+		catch (Throwable ex) {
+			LOGGER.error("Auto update polaris changed value failed, {}", springValue.toString(), ex);
+		}
+	}
+
+
+	/**
+	 * Logic transplanted from DefaultListableBeanFactory.
+	 *
+	 * @see org.springframework.beans.factory.support.DefaultListableBeanFactory#doResolveDependency(org.springframework.beans.factory.config.DependencyDescriptor,
+	 * java.lang.String, java.util.Set, org.springframework.beans.TypeConverter)
+	 */
+	private Object resolvePropertyValue(SpringValue springValue) {
+		// value will never be null
+		Object value = placeholderHelper
+				.resolvePropertyValue(beanFactory, springValue.getBeanName(), springValue.getPlaceholder());
+
+		if (springValue.isJson()) {
+			value = parseJsonValue((String) value, springValue.getTargetType());
+		}
+		else {
+			value = springValue.isField() ? this.typeConverter.convertIfNecessary(value, springValue.getTargetType(), springValue.getField()) :
+					this.typeConverter.convertIfNecessary(value, springValue.getTargetType(),
+							springValue.getMethodParameter());
+		}
+		return value;
+	}
+
+	private Object parseJsonValue(String json, Class<?> targetType) {
+		try {
+			return JacksonUtils.json2JavaBean(json, targetType);
+		}
+		catch (Throwable ex) {
+			LOGGER.error("Parsing json '{}' to type {} failed!", json, targetType, ex);
+			throw ex;
+		}
+	}
+
+	@Override
+	public void setBeanFactory(BeanFactory beanFactory) throws BeansException {
+		this.beanFactory = (ConfigurableBeanFactory) beanFactory;
 	}
 }