+++ spring-cloud-starter-tencent-polaris-config/src/main/java/com/tencent/cloud/polaris/config/config/PolarisConfigProperties.java
@@ -19,6 +19,8 @@ package com.tencent.cloud.polaris.config.config;
 
 import java.util.List;
 
+import com.tencent.cloud.polaris.config.enums.RefreshType;
+
 import org.springframework.boot.context.properties.ConfigurationProperties;
 
 /**
@@ -57,6 +59,11 @@ public class PolarisConfigProperties {
 	private boolean preference = true;
 
 	/**
+	 * Attribute refresh type.
+	 */
+	private RefreshType refreshType = RefreshType.REFRESH_CONTEXT;
+
+	/**
 	 * List of injected configuration files.
 	 */
 	private List<ConfigFileGroup> groups;
@@ -91,6 +98,14 @@ public class PolarisConfigProperties {
 
 	public void setAutoRefresh(boolean autoRefresh) {
 		this.autoRefresh = autoRefresh;
+	}
+
+	public RefreshType getRefreshType() {
+		return refreshType;
+	}
+
+	public void setRefreshType(RefreshType refreshType) {
+		this.refreshType = refreshType;
 	}
 
 	public List<ConfigFileGroup> getGroups() {